backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Azure dependencies
azure-ai-projects>=2.0.0b2
azure-identity==1.15.0
azure-cosmos==4.5.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Async support
aiohttp==3.9.1

# Logging
structlog==24.1.0
